Fix portals cut, Fix probleme when near wall

+++ includes/maths_utils.h
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   math_utils.h                                       :+:      :+:    :+:   */
+/*   maths_utils.h                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: olimarti <olimarti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/01 17:31:00 by olimarti          #+#    #+#             */
-/*   Updated: 2023/11/01 17:40:03 by olimarti         ###   ########.fr       */
+/*   Updated: 2023/11/03 01:39:42 by olimarti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,6 @@ t_point2d	vector4d_to_point2d(t_vector4d *vec4d);
 t_vector4d	point2d_to_vector4d_cpy(t_point2d point);
 t_point2d	vector4d_to_point2d_cpy(t_vector4d vec4d);
 
-int	point_space_partitioning(t_segment_d *separator, t_point2d *point);
+double	point_space_partitioning(t_segment_d *separator, t_point2d *point);
 
 #endif
